[UPDATE] Responsive filter
